refactor: replace local-jazz-server with NAPI TestingServer and native pushSchemaCatalogue
Removes the child-process-based `local-jazz-server.ts` entirely. All test
setup code now uses the in-process NAPI `TestingServer`. Schema catalogue
push is now a native NAPI function backed by a new `schema_catalogue` lib
module in the `jazz-tools` crate, so the CLI binary is no longer required
at test time.

- Extract `jazz_tools::schema_catalogue::push()` into the library crate
  so it can be called from NAPI without spawning a subprocess
- Simplify `commands/schema_push.rs` to delegate to the new lib function
- Add `adminSecret`/`backendSecret` options to `TestingServerBuilder` and
  expose them as getters on `TestingServer` so callers can pass their own
  secrets and read them back
- Add `pushSchemaCatalogue` NAPI function and update `jazz-napi` bindings
- Migrate all consumers (examples, inspector, napi integration tests)
