+++ examples/todo-client-localfirst-svelte/tests/browser/global-setup.ts
@@ -1,14 +1,10 @@
 import { join } from "node:path";
-import {
-  type LocalJazzServerHandle,
-  startLocalJazzServer,
-  pushSchemaCatalogue,
-} from "jazz-tools/testing";
+import { TestingServer, pushSchemaCatalogue } from "jazz-tools/testing";
 import { TEST_PORT, ADMIN_SECRET, APP_ID } from "./test-constants.js";
 
 export { TEST_PORT, ADMIN_SECRET, APP_ID };
 
-let server: Promise<LocalJazzServerHandle> | null = null;
+let server: Promise<TestingServer> | null = null;
 
 export async function setup(): Promise<void> {
   if (server) {
@@ -16,19 +12,18 @@ export async function setup(): Promise<void> {
     return;
   }
 
-  server = startLocalJazzServer({
+  server = TestingServer.start({
     appId: APP_ID,
     port: TEST_PORT,
     adminSecret: ADMIN_SECRET,
-    healthTimeoutMs: 5000,
   });
 
   const serverHandle = await server;
 
   await pushSchemaCatalogue({
     serverUrl: serverHandle.url,
-    appId: APP_ID,
-    adminSecret: ADMIN_SECRET,
+    appId: serverHandle.appId,
+    adminSecret: serverHandle.adminSecret,
     schemaDir: join(import.meta.dirname ?? __dirname, "../../schema"),
   });
 }
