+++ examples/todo-client-localfirst-ts/tests/browser/global-setup.ts
@@ -1,20 +1,19 @@
-import { type LocalJazzServerHandle, startLocalJazzServer } from "jazz-tools/testing";
+import { TestingServer } from "jazz-tools/testing";
 import { TEST_PORT, JWT_SECRET, ADMIN_SECRET, APP_ID } from "./test-constants.js";
 
 export { TEST_PORT, JWT_SECRET, ADMIN_SECRET, APP_ID };
 
-let server: Promise<LocalJazzServerHandle> | null = null;
+let server: Promise<TestingServer> | null = null;
 export async function setup(): Promise<void> {
   if (server) {
     await server;
     return;
   }
 
-  server = startLocalJazzServer({
+  server = TestingServer.start({
     appId: APP_ID,
     port: TEST_PORT,
     adminSecret: ADMIN_SECRET,
-    healthTimeoutMs: 5000,
   });
 
   await server;
