+++ packages/jazz-tools/src/testing/index.ts
@@ -1,8 +1,2 @@
-export {
-  startLocalJazzServer,
-  pushSchemaCatalogue,
-  type LocalJazzServerHandle,
-  type StartLocalJazzServerOptions,
-  type PushSchemaCatalogueOptions,
-} from "./local-jazz-server.js";
+export { TestingServer, pushSchemaCatalogue } from "jazz-napi";
 export { createPolicyTestApp, PolicyTestApp } from "./policy-test-app.js";
